+++ common/core/src/test/java/zingg/common/core/preprocess/TestPreprocessors.java
@@ -14,7 +14,7 @@ import zingg.common.client.*;
 import zingg.common.client.util.DFObjectUtil;
 import zingg.common.core.context.Context;
 import zingg.common.core.data.EventTestData;
-import zingg.common.core.preprocess.model.PriorStopWordProcess;
+import zingg.common.core.preprocess.stopwords.model.PriorStopWordProcess;
 
 public abstract class TestPreprocessors<S,D,R,C,T> {
