+++ common/core/src/test/java/zingg/common/core/preprocess/stopwords/model/Statement.java
@@ -1,4 +1,4 @@
-package zingg.common.core.preprocess.model;
+package zingg.common.core.preprocess.stopwords.model;
 
 public class Statement {
     public final String statement;
